pre-backend validate rework
