+++ front-end/src/utils/validateTest.js
@@ -1,3 +1,63 @@
+//import { today } from "./date-time";
+function asDateString(date) {
+	return `${date.getFullYear().toString(10)}-${(date.getMonth() + 1)
+		.toString(10)
+		.padStart(2, "0")}-${date.getDate().toString(10).padStart(2, "0")}`;
+}
+function today() {
+	return asDateString(new Date());
+}
+function theValidator(formData, setError) {
+	setError(null);
+	let message = "";
+	if (!isTimeOpen(formData.reservation_time)) {
+		message += "  /Closed only open 10:30 AM - 10:30 PM with 1hr window ";
+	}
+	if (isTuesday(formData.reservation_date)) {
+		message += "  /Closed Tuesdays";
+	}
+	if (isPast(formData.reservation_date)) {
+		message += " /Must be in Future";
+	}
+	if (!isTimeValid(formData.reservation_time, formData.reservation_date)) {
+		message += "/ Must be in Future ";
+	}
+	if (message.length) {
+		setError(new Error(message));
+		return false;
+	} else {
+		return true;
+	}
+}
+
+function mobileValidate(phoneNumber, currentLength) {
+	if (phoneNumber[phoneNumber.length - 1] === "-") {
+		phoneNumber = phoneNumber.slice(0, phoneNumber.length - 1);
+	}
+	if (phoneNumber.length === 3) {
+		if (currentLength < phoneNumber.length) {
+			phoneNumber += "-";
+		}
+	}
+
+	if (phoneNumber.length >= 9) {
+		if (!phoneNumber.includes("-")) {
+			const temp1 = phoneNumber.substr(0, 3);
+			const temp2 = phoneNumber.substr(3, 3);
+			const temp3 = phoneNumber.substr(6, 4);
+			phoneNumber = `${temp1}-${temp2}-${temp3}`;
+		} else if (phoneNumber[7] !== "-") {
+			const temp1 = phoneNumber.substr(0, 3);
+			const temp2 = phoneNumber.substr(4, 3);
+			const temp3 = phoneNumber.substr(7, 4);
+			phoneNumber = `${temp1}-${temp2}-${temp3}`;
+		}
+	}
+	if (phoneNumber.length > 12) {
+		phoneNumber = phoneNumber.slice(0, 12);
+	}
+	return phoneNumber;
+}
 function isTuesday(date) {
 	const temp = date.split("-");
 	const newDate = new Date(
@@ -52,4 +112,11 @@ function isTimeOpen(time) {
 	}
 	return false;
 }
-module.exports = { isPast, isTuesday, isTimeValid, isTimeOpen };
+module.exports = {
+	isPast,
+	isTuesday,
+	isTimeValid,
+	isTimeOpen,
+	mobileValidate,
+	theValidator,
+};
